update gmpi generator
