+++ mmedit/models/editors/gmpi/__init__.py
@@ -1,5 +1,4 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .camera import GMPICamera
 from .gmpi_generator import GMPIGenerator
 
-__all__ = ['GMPIGenerator', 'GMPICamera']
+__all__ = ['GMPIGenerator']
