--- mmedit/models/editors/gmpi/gmpi.py
@@ -1,1 +1,0 @@
-# Copyright (c) OpenMMLab. All rights reserved.
